app/requirements.txt
numpy>=1.26
PySide6>=6.7
pyqtgraph>=0.13
pyusb>=1.2
hidapi>=0.14
